==> Programación_de_Inteligencia_Artificial/Pandas/Dataframes/01_dataframes/09_Ejercicio.py <==
# 09. **Crea un gráfico de dispersión que muestre la relación entre las notas finales de 'Programación' y 'Base de Datos'.**
''' Crea un gráfico de dispersión que muestre la relación entre las notas finales de 'Programación' y 'Base de Datos'.'''
#todo
